+++ src/test/java/com/hedvig/notificationService/customerio/events/CreateTmpSignEventTest.kt
@@ -2,7 +2,6 @@ package com.hedvig.notificationService.customerio.events
 
 import com.hedvig.notificationService.customerio.AgreementType
 import com.hedvig.notificationService.customerio.ContractInfo
-import com.hedvig.notificationService.customerio.CustomerioEventCreatorImpl
 import com.hedvig.notificationService.customerio.ProductPricingFacade
 import com.hedvig.notificationService.customerio.state.CustomerioState
 import io.mockk.MockKAnnotations
